Add One step ahead algorithm.
The idea is to compute the expectation of eliminated combinations for
each possible proposition. Then the proposition that has the highest
expectation is played.

On 100 games: Average:5.01 max:6 min:3 in 1546.963572025299 seconds
My exhaustive search shows the real average is closer to 5.06.

More details (in French) at http://colin.pitrat.free.fr/?p=539